requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
google-genai>=0.3.0,<1.0.0
psycopg[binary]>=3.2,<4.0
redis>=5.0,<6.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0

